Update: finished task actions

+++ src/services/taskService.ts
@@ -1,5 +1,5 @@
 import { PrismaClient, type Task } from "@prisma/client";
-import type { TaskBody } from "../interfaces/Task";
+import type { TaskBody, TaskUpdatable } from "../interfaces/Task";
 import ProjectService from "./projectService";
 import UserService from "./userService";
 
@@ -17,15 +17,33 @@ const TaskService = {
     },
 
     findByUser: async (userId: string) => {
-        return prisma.task.findMany({ where: { userId }})
+        return prisma.task.findMany({ where: { userId } })
     },
 
-    edit: async(task: TaskBody) => {},
+    update: async (taskId: string, userId: string, { name, description, status }: TaskUpdatable) => {
+        try {
+            const validStatus = ["pending", "completed", "ongoing"];
+            const updateData: TaskUpdatable = {};
+            const isTaskCreator = await prisma.task.findUnique({ where: { id: taskId, userId: userId } });
+            if (!isTaskCreator) {
+                throw new Error("Unauthorized! cannot update task");
+            }
+            if (name) updateData.name = name;
+            if (description) updateData.description = description;
+            if (status) {
+                const canUpdateStatus = validStatus.indexOf(status) > -1;
+                if (canUpdateStatus) updateData.status = status;
+            }
+            return prisma.task.update({ where: { id: taskId }, data: updateData });
+        } catch (error) {
+            throw error;
+        }
+    },
 
-    remove: async(taskId: string) => {
-        return prisma.task.delete({ where: { id: taskId }});
+    remove: async (taskId: string, userId: string) => {
+        return prisma.task.delete({ where: { id: taskId, userId: userId } });
     }
 }
 
 
-export default TaskService;
+export default TaskService;
